Add smart view class
